requirements.txt
python-dotenv~=0.21.1
playwright~=1.30.0